+++ src/digest.rs
@@ -1,5 +1,7 @@
 //! Digest-able byte sequence. See [`Digest`].
 
+use std::slice::SliceIndex;
+
 /// A digest-able byte sequence.
 ///
 /// Built-in implementations are provided for `[u8]` and [`str`].
@@ -14,17 +16,13 @@ pub trait Digest {
   /// Convert self to a byte slice.
   fn as_bytes(&self) -> &[u8];
 
-  /// Digest the next `n` bytes. Return the rest.
-  /// # Safety
-  /// You should ensure that `n` is valid according to [`Digest::validate`].
-  /// This will be checked using [`debug_assert!`].
-  unsafe fn digest_unchecked(&self, n: usize) -> &Self;
+  /// Get a subslice of `self` if it is valid.
+  fn get<I: SliceIndex<Self>>(&self, i: I) -> Option<&I::Output>;
 
-  /// Return the first `n` bytes.
+  /// Get an unchecked subslice of `self` without bound checking.
   /// # Safety
   /// You should ensure that `n` is valid according to [`Digest::validate`].
-  /// This will be checked using [`debug_assert!`].
-  unsafe fn span_unchecked(&self, n: usize) -> &Self;
+  unsafe fn get_unchecked<I: SliceIndex<Self>>(&self, i: I) -> &I::Output;
 }
 
 impl Digest for [u8] {
@@ -39,15 +37,13 @@ impl Digest for [u8] {
   }
 
   #[inline]
-  unsafe fn digest_unchecked(&self, n: usize) -> &Self {
-    debug_assert!(self.validate(n));
-    self.get_unchecked(n..)
+  fn get<I: SliceIndex<Self>>(&self, i: I) -> Option<&I::Output> {
+    self.get(i)
   }
 
   #[inline]
-  unsafe fn span_unchecked(&self, n: usize) -> &Self {
-    debug_assert!(self.validate(n));
-    self.get_unchecked(..n)
+  unsafe fn get_unchecked<I: SliceIndex<Self>>(&self, i: I) -> &I::Output {
+    self.get_unchecked(i)
   }
 }
 
@@ -59,19 +55,17 @@ impl Digest for str {
 
   #[inline]
   fn as_bytes(&self) -> &[u8] {
-    <str>::as_bytes(self)
+    self.as_bytes()
   }
 
   #[inline]
-  unsafe fn digest_unchecked(&self, n: usize) -> &Self {
-    debug_assert!(self.validate(n));
-    self.get_unchecked(n..)
+  fn get<I: SliceIndex<Self>>(&self, i: I) -> Option<&I::Output> {
+    self.get(i)
   }
 
   #[inline]
-  unsafe fn span_unchecked(&self, n: usize) -> &Self {
-    debug_assert!(self.validate(n));
-    self.get_unchecked(..n)
+  unsafe fn get_unchecked<I: SliceIndex<Self>>(&self, i: I) -> &I::Output {
+    self.get_unchecked(i)
   }
 }
 
@@ -87,73 +81,26 @@ mod tests {
     assert!(bytes.validate(2));
     assert!(bytes.validate(3));
     assert!(!bytes.validate(4));
-    assert_eq!(unsafe { bytes.digest_unchecked(0) }, b"123");
-    assert_eq!(unsafe { bytes.digest_unchecked(1) }, b"23");
-    assert_eq!(unsafe { bytes.digest_unchecked(2) }, b"3");
-    assert_eq!(unsafe { bytes.digest_unchecked(3) }, b"");
-    assert_eq!(unsafe { bytes.span_unchecked(0) }, b"");
-    assert_eq!(unsafe { bytes.span_unchecked(1) }, b"1");
-    assert_eq!(unsafe { bytes.span_unchecked(2) }, b"12");
-    assert_eq!(unsafe { bytes.span_unchecked(3) }, b"123");
-  }
-
-  #[test]
-  #[should_panic]
-  fn digest_bytes_overflow() {
-    let bytes = b"123" as &[u8];
-    unsafe { bytes.digest_unchecked(4) };
-  }
-
-  #[test]
-  #[should_panic]
-  fn digest_bytes_span_overflow() {
-    let bytes = b"123" as &[u8];
-    unsafe { bytes.span_unchecked(4) };
+    assert_eq!(bytes.as_bytes(), b"123");
+    assert_eq!(<[u8] as Digest>::get(bytes, 0), Some(&b'1'));
+    assert_eq!(<[u8] as Digest>::get(bytes, 0..), Some(b"123" as &[u8]));
+    assert_eq!(unsafe { <[u8] as Digest>::get_unchecked(bytes, 0) }, &b'1');
+    assert_eq!(
+      unsafe { <[u8] as Digest>::get_unchecked(bytes, 0..) },
+      b"123"
+    );
   }
 
   #[test]
   fn digest_str() {
-    let text = "123";
+    let text = "好";
     assert!(text.validate(0));
-    assert!(text.validate(1));
-    assert!(text.validate(2));
+    assert!(!text.validate(1));
+    assert!(!text.validate(2));
     assert!(text.validate(3));
     assert!(!text.validate(4));
-    assert_eq!(unsafe { text.digest_unchecked(0) }, "123");
-    assert_eq!(unsafe { text.digest_unchecked(1) }, "23");
-    assert_eq!(unsafe { text.digest_unchecked(2) }, "3");
-    assert_eq!(unsafe { text.digest_unchecked(3) }, "");
-    assert_eq!(unsafe { text.span_unchecked(0) }, "");
-    assert_eq!(unsafe { text.span_unchecked(1) }, "1");
-    assert_eq!(unsafe { text.span_unchecked(2) }, "12");
-    assert_eq!(unsafe { text.span_unchecked(3) }, "123");
-  }
-
-  #[test]
-  #[should_panic]
-  fn digest_str_invalid_code_point() {
-    let text = "好";
-    unsafe { text.digest_unchecked(1) };
-  }
-
-  #[test]
-  #[should_panic]
-  fn digest_str_span_invalid_code_point() {
-    let text = "好";
-    unsafe { text.span_unchecked(1) };
-  }
-
-  #[test]
-  #[should_panic]
-  fn digest_str_overflow() {
-    let text = "123";
-    unsafe { text.digest_unchecked(4) };
-  }
-
-  #[test]
-  #[should_panic]
-  fn digest_str_span_overflow() {
-    let text = "123";
-    unsafe { text.span_unchecked(4) };
+    assert_eq!(<str as Digest>::as_bytes(text), [229, 165, 189]);
+    assert_eq!(<str as Digest>::get(text, 0..), Some("好"));
+    assert_eq!(unsafe { <str as Digest>::get_unchecked(text, 0..) }, "好");
   }
 }